Make flagged review async
